Split controller test class into partial classes
