+++ src/web/EmployerFavouritesWeb/Controllers/ApprenticeshipsController.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
+﻿using System.Linq;
 using System.Threading.Tasks;
 using DfE.EmployerFavourites.Application.Commands;
 using DfE.EmployerFavourites.Application.Queries;
