+++ src/web/IntegrationTests/IntegrationTests.csproj
@@ -28,4 +28,7 @@
     </Content>   
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Tests\" />
+  </ItemGroup>
 </Project>
